move verrazzano and module controller
