+++ platform-operator/controllers/module/component-handler/common/handler_utils.go
