+++ platform-operator/controllers/module/component-handler/common/vzcomponent_status.go
