+++ platform-operator/controllers/module/component-handler/installupdate/install_update_handler.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package upgrade
+package installupdate
 
 import (
 	moduleapi "github.com/verrazzano/verrazzano-modules/module-operator/apis/platform/v1alpha1"
@@ -11,106 +11,115 @@ import (
 	vzerrors "github.com/verrazzano/verrazzano/pkg/controller/errors"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
-	"github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/module/component-handler/common"
+	common2 "github.com/verrazzano/verrazzano/platform-operator/controllers/module/component-handler/common"
 )
 
-type ComponentHandler struct{}
+type ActionType string
+
+const (
+	InstallAction ActionType = constants.InstallOperation
+	UpdateAction  ActionType = constants.UpdateOperation
+)
+
+type ComponentHandler struct {
+	action ActionType
+}
 
 var (
 	_ handlerspi.StateMachineHandler = &ComponentHandler{}
 )
 
-func NewHandler() handlerspi.StateMachineHandler {
-	return &ComponentHandler{}
+func NewHandler(action ActionType) handlerspi.StateMachineHandler {
+	return &ComponentHandler{action: action}
 }
 
 // GetWorkName returns the work name
 func (h ComponentHandler) GetWorkName() string {
-	return "upgrade"
+	return string(h.action)
 }
 
-// IsWorkNeeded returns true if upgrade is needed
+// IsWorkNeeded returns true if install/update is needed
 func (h ComponentHandler) IsWorkNeeded(ctx handlerspi.HandlerContext) (bool, result.Result) {
-	module := ctx.CR.(*moduleapi.Module)
-	needUpgrade := module.Spec.Version != "" && (module.Spec.Version != module.Status.LastSuccessfulVersion)
-	return needUpgrade, result.NewResult()
+	return true, result.NewResult()
 }
 
 // CheckDependencies checks if the dependencies are ready
 func (h ComponentHandler) CheckDependencies(ctx handlerspi.HandlerContext) result.Result {
-	return common.CheckDependencies(ctx, string(constants.UpgradeOperation), moduleapi.ReadyReasonUpgradeStarted)
+	return common2.CheckDependencies(ctx, string(h.action), h.getStartedReason())
 }
 
-// PreWorkUpdateStatus updates the status for the pre-work state
+// PreWorkUpdateStatus does the pre-Work status update
 func (h ComponentHandler) PreWorkUpdateStatus(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
 	// Update the Verrazzano component status
-	nsn, err := common.GetVerrazzanoNSN(ctx)
+	nsn, err := common2.GetVerrazzanoNSN(ctx)
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
-	sd := common.StatusData{
+	sd := common2.StatusData{
 		Vznsn:    *nsn,
-		CondType: vzapi.CondUpgradeStarted,
+		CondType: vzapi.CondInstallStarted,
 		CompName: module.Spec.ModuleName,
-		Msg:      string(vzapi.CondUpgradeStarted),
+		Msg:      string(vzapi.CondInstallStarted),
 		Ready:    false,
 	}
-	res := common.UpdateVerrazzanoComponentStatus(ctx, sd)
+	res := common2.UpdateVerrazzanoComponentStatus(ctx, sd)
 	if res.ShouldRequeue() {
 		return res
 	}
 
 	// Update the module status
-	return modulestatus.UpdateReadyConditionReconciling(ctx, module, moduleapi.ReadyReasonUpgradeStarted)
+	return modulestatus.UpdateReadyConditionReconciling(ctx, module, h.getStartedReason())
 }
 
 // PreWork does the pre-work
 func (h ComponentHandler) PreWork(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
-	compCtx, comp, err := common.GetComponentAndContext(ctx, constants.UpgradeOperation)
+	compCtx, comp, err := common2.GetComponentAndContext(ctx, string(h.action))
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 
-	// Do the pre-upgrade
-	if err := comp.PreUpgrade(compCtx); err != nil {
+	// Do the pre-install
+	if err := comp.PreInstall(compCtx); err != nil {
 		if !vzerrors.IsRetryableError(err) {
-			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUpgradeStarted, err.Error())
+			modulestatus.UpdateReadyConditionFailed(ctx, module, h.getStartedReason(), err.Error())
 		}
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
+	modulestatus.UpdateReadyConditionReconciling(ctx, module, h.getStartedReason())
 	return result.NewResult()
 }
 
-// DoWorkUpdateStatus updates the status for the work state
+// DoWorkUpdateStatus does th status update
 func (h ComponentHandler) DoWorkUpdateStatus(ctx handlerspi.HandlerContext) result.Result {
 	return result.NewResult()
 }
 
-// DoWork upgrades the module using Helm
+// DoWork installs the module using Helm
 func (h ComponentHandler) DoWork(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
-	compCtx, comp, err := common.GetComponentAndContext(ctx, constants.UpgradeOperation)
+	compCtx, comp, err := common2.GetComponentAndContext(ctx, string(h.action))
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 
-	if err := comp.Upgrade(compCtx); err != nil {
+	if err := comp.Install(compCtx); err != nil {
 		if !vzerrors.IsRetryableError(err) {
-			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUpgradeStarted, err.Error())
+			modulestatus.UpdateReadyConditionFailed(ctx, module, h.getStartedReason(), err.Error())
 		}
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
+	modulestatus.UpdateReadyConditionReconciling(ctx, module, h.getStartedReason())
 	return result.NewResult()
 }
 
-// IsWorkDone indicates whether a module is upgraded and ready
+// IsWorkDone Indicates whether a module is installed and ready
 func (h ComponentHandler) IsWorkDone(ctx handlerspi.HandlerContext) (bool, result.Result) {
-	compCtx, comp, err := common.GetComponentAndContext(ctx, constants.UpgradeOperation)
+	compCtx, comp, err := common2.GetComponentAndContext(ctx, string(h.action))
 	if err != nil {
 		return false, result.NewResultShortRequeueDelayWithError(err)
 	}
@@ -124,45 +133,64 @@ func (h ComponentHandler) PostWorkUpdateStatus(ctx handlerspi.HandlerContext) re
 	return result.NewResult()
 }
 
-// PostWork does installation pre-work
+// PostWork does installation post-work
 func (h ComponentHandler) PostWork(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
-	compCtx, comp, err := common.GetComponentAndContext(ctx, constants.UpgradeOperation)
+	compCtx, comp, err := common2.GetComponentAndContext(ctx, string(h.action))
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
-	if err := comp.PostUpgrade(compCtx); err != nil {
+	if err := comp.PostInstall(compCtx); err != nil {
 		if !vzerrors.IsRetryableError(err) {
-			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUpgradeStarted, err.Error())
+			modulestatus.UpdateReadyConditionFailed(ctx, module, h.getStartedReason(), err.Error())
 		}
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
+	modulestatus.UpdateReadyConditionReconciling(ctx, module, h.getStartedReason())
 	return result.NewResult()
 }
 
 // WorkCompletedUpdateStatus updates the status to completed
 func (h ComponentHandler) WorkCompletedUpdateStatus(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
+	var reason moduleapi.ModuleConditionReason
+	var cond vzapi.ConditionType
+
+	if h.action == InstallAction {
+		reason = moduleapi.ReadyReasonInstallSucceeded
+		cond = vzapi.CondInstallComplete
+	} else {
+		reason = moduleapi.ReadyReasonUpdateSucceeded
+		// Note, Verrazzano uses install condition for update
+		cond = vzapi.CondInstallComplete
+	}
 
 	// Update the Verrazzano component status
-	nsn, err := common.GetVerrazzanoNSN(ctx)
+	nsn, err := common2.GetVerrazzanoNSN(ctx)
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
-	sd := common.StatusData{
+	sd := common2.StatusData{
 		Vznsn:       *nsn,
-		CondType:    vzapi.CondUpgradeComplete,
+		CondType:    cond,
 		CompName:    module.Spec.ModuleName,
 		CompVersion: module.Spec.Version,
-		Msg:         string(vzapi.CondUpgradeComplete),
+		Msg:         string(cond),
 		Ready:       true,
 	}
-	res := common.UpdateVerrazzanoComponentStatus(ctx, sd)
+	res := common2.UpdateVerrazzanoComponentStatus(ctx, sd)
 	if res.ShouldRequeue() {
 		return res
 	}
 
 	// Update the module status
-	return modulestatus.UpdateReadyConditionSucceeded(ctx, module, moduleapi.ReadyReasonUpgradeSucceeded)
+	return modulestatus.UpdateReadyConditionSucceeded(ctx, module, reason)
+}
+
+func (h ComponentHandler) getStartedReason() moduleapi.ModuleConditionReason {
+	if h.action == InstallAction {
+		return moduleapi.ReadyReasonInstallStarted
+	}
+	return moduleapi.ReadyReasonUpdateStarted
 }
