+++ platform-operator/controllers/module/component-handler/upgrade/upgrade_handler.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package delete
+package upgrade
 
 import (
 	moduleapi "github.com/verrazzano/verrazzano-modules/module-operator/apis/platform/v1alpha1"
@@ -11,7 +11,7 @@ import (
 	vzerrors "github.com/verrazzano/verrazzano/pkg/controller/errors"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
-	"github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/module/component-handler/common"
+	common2 "github.com/verrazzano/verrazzano/platform-operator/controllers/module/component-handler/common"
 )
 
 type ComponentHandler struct{}
@@ -26,101 +26,97 @@ func NewHandler() handlerspi.StateMachineHandler {
 
 // GetWorkName returns the work name
 func (h ComponentHandler) GetWorkName() string {
-	return "uninstall"
+	return "upgrade"
 }
 
-// IsWorkNeeded returns true if uninstall is needed
+// IsWorkNeeded returns true if upgrade is needed
 func (h ComponentHandler) IsWorkNeeded(ctx handlerspi.HandlerContext) (bool, result.Result) {
-	// Always return true so that the post-uninstall can run in the case that the VPO
-	// was restarted
-	return true, result.NewResult()
+	module := ctx.CR.(*moduleapi.Module)
+	needUpgrade := module.Spec.Version != "" && (module.Spec.Version != module.Status.LastSuccessfulVersion)
+	return needUpgrade, result.NewResult()
 }
 
 // CheckDependencies checks if the dependencies are ready
 func (h ComponentHandler) CheckDependencies(ctx handlerspi.HandlerContext) result.Result {
-	return result.NewResult()
+	return common2.CheckDependencies(ctx, string(constants.UpgradeOperation), moduleapi.ReadyReasonUpgradeStarted)
 }
 
-// PreWorkUpdateStatus does the lifecycle pre-Work status update
+// PreWorkUpdateStatus updates the status for the pre-work state
 func (h ComponentHandler) PreWorkUpdateStatus(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
 	// Update the Verrazzano component status
-	nsn, err := common.GetVerrazzanoNSN(ctx)
+	nsn, err := common2.GetVerrazzanoNSN(ctx)
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
-	sd := common.StatusData{
+	sd := common2.StatusData{
 		Vznsn:    *nsn,
-		CondType: vzapi.CondUninstallStarted,
+		CondType: vzapi.CondUpgradeStarted,
 		CompName: module.Spec.ModuleName,
-		Msg:      string(vzapi.CondUninstallStarted),
+		Msg:      string(vzapi.CondUpgradeStarted),
 		Ready:    false,
 	}
-	res := common.UpdateVerrazzanoComponentStatus(ctx, sd)
+	res := common2.UpdateVerrazzanoComponentStatus(ctx, sd)
 	if res.ShouldRequeue() {
 		return res
 	}
 
 	// Update the module status
-	return modulestatus.UpdateReadyConditionReconciling(ctx, module, moduleapi.ReadyReasonUninstallStarted)
+	return modulestatus.UpdateReadyConditionReconciling(ctx, module, moduleapi.ReadyReasonUpgradeStarted)
 }
 
 // PreWork does the pre-work
 func (h ComponentHandler) PreWork(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
-	compCtx, comp, err := common.GetComponentAndContext(ctx, constants.UninstallOperation)
+	compCtx, comp, err := common2.GetComponentAndContext(ctx, constants.UpgradeOperation)
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 
-	// Do the pre-delete
-	if err := comp.PreUninstall(compCtx); err != nil {
+	// Do the pre-upgrade
+	if err := comp.PreUpgrade(compCtx); err != nil {
 		if !vzerrors.IsRetryableError(err) {
-			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUninstallStarted, err.Error())
+			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUpgradeStarted, err.Error())
 		}
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 	return result.NewResult()
 }
 
-// DoWorkUpdateStatus does the work status update
+// DoWorkUpdateStatus updates the status for the work state
 func (h ComponentHandler) DoWorkUpdateStatus(ctx handlerspi.HandlerContext) result.Result {
 	return result.NewResult()
 }
 
-// DoWork uninstalls the module using Helm
+// DoWork upgrades the module using Helm
 func (h ComponentHandler) DoWork(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
-	compCtx, comp, err := common.GetComponentAndContext(ctx, constants.UninstallOperation)
+	compCtx, comp, err := common2.GetComponentAndContext(ctx, constants.UpgradeOperation)
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 
-	if err := comp.Uninstall(compCtx); err != nil {
+	if err := comp.Upgrade(compCtx); err != nil {
 		if !vzerrors.IsRetryableError(err) {
-			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUninstallStarted, err.Error())
+			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUpgradeStarted, err.Error())
 		}
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 	return result.NewResult()
 }
 
-// IsWorkDone Indicates whether a module is uninstalled
+// IsWorkDone indicates whether a module is upgraded and ready
 func (h ComponentHandler) IsWorkDone(ctx handlerspi.HandlerContext) (bool, result.Result) {
-	compCtx, comp, err := common.GetComponentAndContext(ctx, constants.UninstallOperation)
+	compCtx, comp, err := common2.GetComponentAndContext(ctx, constants.UpgradeOperation)
 	if err != nil {
 		return false, result.NewResultShortRequeueDelayWithError(err)
 	}
 
-	exists, err := comp.Exists(compCtx)
-	if err != nil {
-		ctx.Log.ErrorfThrottled("Error checking if Helm release exists for %s/%s", ctx.HelmRelease.Namespace, ctx.HelmRelease.Name)
-		return false, result.NewResultShortRequeueDelayWithError(err)
-	}
-	return !exists, result.NewResult()
+	ready := comp.IsReady(compCtx)
+	return ready, result.NewResult()
 }
 
 // PostWorkUpdateStatus does the post-work status update
@@ -132,34 +128,41 @@ func (h ComponentHandler) PostWorkUpdateStatus(ctx handlerspi.HandlerContext) re
 func (h ComponentHandler) PostWork(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
-	compCtx, comp, err := common.GetComponentAndContext(ctx, constants.UninstallOperation)
+	compCtx, comp, err := common2.GetComponentAndContext(ctx, constants.UpgradeOperation)
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
-	if err := comp.PostUninstall(compCtx); err != nil {
+	if err := comp.PostUpgrade(compCtx); err != nil {
 		if !vzerrors.IsRetryableError(err) {
-			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUninstallStarted, err.Error())
+			modulestatus.UpdateReadyConditionFailed(ctx, module, moduleapi.ReadyReasonUpgradeStarted, err.Error())
 		}
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 	return result.NewResult()
 }
 
-// WorkCompletedUpdateStatus does the lifecycle completed Work status update
+// WorkCompletedUpdateStatus updates the status to completed
 func (h ComponentHandler) WorkCompletedUpdateStatus(ctx handlerspi.HandlerContext) result.Result {
 	module := ctx.CR.(*moduleapi.Module)
 
-	// Update the Verrazzano component status to disabled
-	vzNSN, err := common.GetVerrazzanoNSN(ctx)
+	// Update the Verrazzano component status
+	nsn, err := common2.GetVerrazzanoNSN(ctx)
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
-
-	res := common.UpdateVerrazzanoComponentStatusToDisabled(ctx, *vzNSN, module.Spec.ModuleName)
+	sd := common2.StatusData{
+		Vznsn:       *nsn,
+		CondType:    vzapi.CondUpgradeComplete,
+		CompName:    module.Spec.ModuleName,
+		CompVersion: module.Spec.Version,
+		Msg:         string(vzapi.CondUpgradeComplete),
+		Ready:       true,
+	}
+	res := common2.UpdateVerrazzanoComponentStatus(ctx, sd)
 	if res.ShouldRequeue() {
 		return res
 	}
 
 	// Update the module status
-	return modulestatus.UpdateReadyConditionSucceeded(ctx, module, moduleapi.ReadyReasonUninstallSucceeded)
+	return modulestatus.UpdateReadyConditionSucceeded(ctx, module, moduleapi.ReadyReasonUpgradeSucceeded)
 }
