+++ platform-operator/controllers/module/doc.go
