+++ platform-operator/controllers/verrazzano/controller/custom/rancher.go
