+++ platform-operator/controllers/verrazzano/controller/finalizer.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package verrazzano
+package controller
 
 import (
 	"context"
@@ -12,8 +12,8 @@ import (
 	vzv1alpha1 "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/rancher"
 	componentspi "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
+	custom2 "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/controller/custom"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/transform"
-	"github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/verrazzano/custom"
 	"go.uber.org/zap"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -103,7 +103,7 @@ func (r Reconciler) PostRemoveFinalizer(controllerCtx controllerspi.ReconcileCon
 
 // preUninstall does all the global preUninstall
 func (r Reconciler) preUninstall(log vzlog.VerrazzanoLogger, actualCR *vzv1alpha1.Verrazzano, effectiveCR *vzv1alpha1.Verrazzano) result.Result {
-	if res := custom.PreUninstallRancher(r.Client, log, actualCR, effectiveCR); res.ShouldRequeue() {
+	if res := custom2.PreUninstallRancher(r.Client, log, actualCR, effectiveCR); res.ShouldRequeue() {
 		return res
 	}
 
@@ -146,7 +146,7 @@ func (r Reconciler) preUninstallMC(log vzlog.VerrazzanoLogger, actualCR *vzv1alp
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
-	if err := custom.DeleteMCResources(componentCtx); err != nil {
+	if err := custom2.DeleteMCResources(componentCtx); err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 
@@ -161,12 +161,12 @@ func (r Reconciler) postUninstallCleanup(log vzlog.VerrazzanoLogger, componentCt
 	}
 
 	log.Once("Global post-uninstall: deleting Istio CA Root Cert")
-	if err := custom.DeleteIstioCARootCert(componentCtx); err != nil {
+	if err := custom2.DeleteIstioCARootCert(componentCtx); err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 
 	log.Once("Global post-uninstall: node exporter cleanup")
-	if err := custom.NodeExporterCleanup(componentCtx.Client(), componentCtx.Log()); err != nil {
+	if err := custom2.NodeExporterCleanup(componentCtx.Client(), componentCtx.Log()); err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 
@@ -175,10 +175,10 @@ func (r Reconciler) postUninstallCleanup(log vzlog.VerrazzanoLogger, componentCt
 	// installed explicitly.
 	if !rancherProvisioned {
 		log.Once("Global post-uninstall: running Rancher cleanup")
-		if err := custom.RunRancherPostUninstall(componentCtx); err != nil {
+		if err := custom2.RunRancherPostUninstall(componentCtx); err != nil {
 			return result.NewResultShortRequeueDelayWithError(err)
 		}
 	}
 	log.Once("Global post-uninstall: deleting namespaces")
-	return custom.DeleteNamespaces(componentCtx, rancherProvisioned)
+	return custom2.DeleteNamespaces(componentCtx, rancherProvisioned)
 }
