+++ platform-operator/controllers/verrazzano/controller/reconciler.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package verrazzano
+package controller
 
 import (
 	"context"
@@ -19,9 +19,9 @@ import (
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/mysql"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/rancher"
 	componentspi "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
+	custom2 "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/controller/custom"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/restart"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/transform"
-	"github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/verrazzano/custom"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/k8s/namespace"
 	"go.uber.org/zap"
@@ -116,13 +116,13 @@ func (r Reconciler) preWork(log vzlog.VerrazzanoLogger, actualCR *vzv1alpha1.Ver
 	}
 
 	// Delete leftover MySQL backup job if we find one.
-	if err := custom.CleanupMysqlBackupJob(log, r.Client); err != nil {
+	if err := custom2.CleanupMysqlBackupJob(log, r.Client); err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
 
 	// if an OCI DNS installation, make sure the secret required exists before proceeding
 	if actualCR.Spec.Components.DNS != nil && actualCR.Spec.Components.DNS.OCI != nil {
-		err := custom.DoesOCIDNSConfigSecretExist(r.Client, actualCR)
+		err := custom2.DoesOCIDNSConfigSecretExist(r.Client, actualCR)
 		if err != nil {
 			return result.NewResultShortRequeueDelayWithError(err)
 		}
@@ -130,7 +130,7 @@ func (r Reconciler) preWork(log vzlog.VerrazzanoLogger, actualCR *vzv1alpha1.Ver
 
 	// Sync the local cluster registration secret that allows the use of MC xyz resources on the
 	// admin cluster without needing a VMC.
-	if err := custom.SyncLocalRegistrationSecret(r.Client); err != nil {
+	if err := custom2.SyncLocalRegistrationSecret(r.Client); err != nil {
 		log.Errorf("Failed to sync the local registration secret: %v", err)
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
@@ -140,7 +140,7 @@ func (r Reconciler) preWork(log vzlog.VerrazzanoLogger, actualCR *vzv1alpha1.Ver
 	if err != nil {
 		return result.NewResultShortRequeueDelayWithError(err)
 	}
-	custom.CreateRancherIngressAndCertCopies(componentCtx)
+	custom2.CreateRancherIngressAndCertCopies(componentCtx)
 
 	return result.NewResult()
 }
