+++ platform-operator/controllers/verrazzano/controller/testdata/test_bom.json
